# packages/skyhelper/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shards-embed": {
      "get": {
        "description": "Retrieves Discord embed data for shard information on a specific date",
        "operationId": "AppController_getShardsEmbed",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Language locale",
            "schema": {
              "example": "en-US",
              "type": "string"
            }
          },
          {
            "name": "user",
            "required": false,
            "in": "query",
            "description": "User ID for personalized content",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noBtn",
            "required": false,
            "in": "query",
            "description": "Exclude buttons from embed",
            "schema": {
              "enum": ["true", "false"],
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "example": "2024-01-15",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discord embed object for shard information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "color": {
                      "type": "number"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get shards embed data",
        "tags": ["Application"]
      }
    },
    "/times-embed": {
      "get": {
        "description": "Retrieves Discord embed data for Sky game event times",
        "operationId": "AppController_getTimesEmbed",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Language locale",
            "schema": {
              "example": "en-US",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discord embed object for event times",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "color": {
                      "type": "number"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get event times embed data",
        "tags": ["Application"]
      }
    },
...
  },
  "info": {
    "title": "SkyHelper API",
    "description": "\n      The SkyHelper Bot API provides comprehensive access to Sky: Children of the Light game data and bot management features.\n      \n      ## Features\n      - **Guild Management**: Configure bot settings for Discord servers\n      - **Game Data**: Access shards, events, traveling spirits, and daily quests information\n      - **User Preferences**: Manage user settings and linked roles\n      - **Live Updates**: Configure automated updates for game events\n      - **Reminders**: Set up notifications for game activities\n      \n      ## Authentication\n      Most endpoints require authentication via Bearer token obtained through Discord OAuth2.\n      The token should be included in the Authorization header as: `Bearer <token>`\n      \n      ## Rate Limiting\n      API requests are rate-limited to prevent abuse. Please respect the limits and implement proper retry logic.\n      \n      ## Support\n      For API support, join our Discord server or check the GitHub repository.\n    ",
    "version": "7.7.0",
    "contact": {
      "name": "SkyHelper Team",
      "url": "https://github.com/imnaiyar/SkyHelper",
      "email": "support@skyhelperbot.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/imnaiyar/SkyHelper/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://api.skyhelperbot.com",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your Discord OAuth2 access token",
        "in": "header"
      }
    },
...
  }
}